+++ src/components/game/Game.tsx
@@ -8,27 +8,9 @@ import FilesSelector, { File } from '../filesSelector/FilesSelector'
 import Sketch from './Sketch'
 import $ from './Game.module.scss'
 import { gameInstance } from './GameStates'
+import { GameProject, buildGameProject, mainProgram } from './gameProject'
 
 const natives = wre as Natives
-const WOLLOK_FILE_EXTENSION = 'wlk'
-const WOLLOK_PROGRAM_EXTENSION = 'wpgm'
-const EXPECTED_WOLLOK_EXTENSIONS = [WOLLOK_FILE_EXTENSION, WOLLOK_PROGRAM_EXTENSION]
-const VALID_IMAGE_EXTENSIONS = ['jpg', 'jpeg', 'png', 'gif']
-export const DEFAULT_GAME_ASSETS_DIR = 'https://raw.githubusercontent.com/uqbar-project/wollok/dev/org.uqbar.project.wollok.game/assets/'
-
-type SourceFile = {
-  name: string
-  content: string
-}
-
-export interface GameProject {
-  main: string
-  sources: SourceFile[]
-  description: string
-  assetsDir: string
-  images: { path: string, imageUrl: string }[]
-}
-
 
 export type GameProps = RouteComponentProps
 const Game = (_: GameProps) => {
@@ -38,11 +20,9 @@ const Game = (_: GameProps) => {
   const loadGame = (files: File[]) => {
     const project = buildGameProject(files)
     const environment = buildEnvironment(project.sources)
-    const programWollokFile = environment.getNodeByFQN<'Package'>(`${project.main}`)
-    const mainWollokProgramName = programWollokFile.members.find(entity => entity.is('Program'))?.name
     const { buildEvaluation, runProgram } = interpret(environment, natives)
     const cleanEval = buildEvaluation()
-    runProgram(`${project.main}.${mainWollokProgramName}`, cleanEval)
+    runProgram(mainProgram(project, environment), cleanEval)
     setGame(project)
     setEvaluation(cleanEval)
   }
@@ -66,29 +46,3 @@ const Game = (_: GameProps) => {
 }
 
 export default memo(Game)
-
-function buildGameProject(files: File[]): GameProject {
-  const repoUri = new URLSearchParams(document.location.search).get('git')?.replace('https://github.com/', '') //TODO: Arreglar las imágenes para evitar esto, sino no va a andar con proyectos locales
-
-  const wpgmFile = files.find(withExtension(WOLLOK_PROGRAM_EXTENSION))
-  if (!wpgmFile) throw new Error('Program not found')
-  const main = wpgmFile.name.replace(`.${WOLLOK_PROGRAM_EXTENSION}`, '').replace(/\//gi, '.')
-  const sources = files.filter(withExtension(...EXPECTED_WOLLOK_EXTENSIONS)).map(({ name, content }) => ({ name, content: content.toString('utf8') }))
-  const assetSource = `https://raw.githubusercontent.com/${repoUri}/master/`
-  const imgFiles = files.filter(withExtension(...VALID_IMAGE_EXTENSIONS))
-  const gameAssetsPaths = imgFiles.map(({ name }) => assetSource + name)
-  const assetFolderName = gameAssetsPaths[0]?.substring(assetSource.length).split('/')[0]
-  const assetsDir = assetSource + assetFolderName + '/'
-  const description = files.find(withExtension('md'))?.content.toString('utf8') || '## No description found'
-  const images = imgFiles.map(({ name, content }) => (
-    {
-      path: name.split('/').pop()!, //TODO: Contemplar sub-directorios
-      imageUrl: URL.createObjectURL(new Blob([content], { type: 'image/png' })),
-    }
-  ))
-
-  return { main, sources, description, assetsDir, images }
-}
-
-const withExtension = (...extensions: string[]) => ({ name }: File) =>
-  extensions.some(extension => name.endsWith(`.${extension}`))
